palestraintuito_bot_finale/requirements.txt
python-telegram-bot==13.15
python-dotenv
requests
apscheduler==3.10.4
